Adding felshed out role ready for CI testing
Adds initial features including:
* Installing postgres from non-system or system-only sources
* Configuring postgres to allow remote connections
* Configuring the system firewall to allow remote connections

Due to differences between packages between operating systems this role requires a number of tasks which appear unnecessary (or which could be condensed) but which are necessary to overcome differences in where configuration files are located for example.

A full set of tests (which pass locally) are included for all major features (including creating roles and databases). Documentation has also been added to cover the specific features of this role, as well as setting a wider context to give a overview of the goals of this role.