Switching to MD5 authentication for non-socket local connections
This changes the authentication method used for connections made to local interfaces (i.e. 127.0.0.1) to use password (MD5) authentication over 'ident'.

This is to support using scripts (which connect using a username and password) on machines with a local PostgreSQL instance - which will be typical in development environments.

Using 'ident' requires the use of an operating system user, which may either not exist for a given database role (user), is unavailable to the calling script, or is an unsupported method of operation within the calling script.

This change does not change the authentication method used for local socket connections (which remains as 'peer') or remote connections (which remains as 'MD5').

As this change only affects local connections it is to considered to be a significant change in terms of security, and the 'ident' authentication method would likely not have been used for these types of connection, as typically the local socket would be used, rather than connecting locally over TCP.

Therefore this change is suitable to include in a minor role release.