Adding detection for failed database initialisations
On CentOS the database needs to be manually initiated, which can in some cases fail.

To workaround this, this role will delete the database installation where no postgres configuration file is present (which should only occur in error conditions).

If there is no configuration file, the data directory (which will typically only contain a log file) will be removed, allowing the subsequent 'init' command to succeed.